visiocraft_ai/requirements.txt
openai
flask
python-dotenv
google-generativeai